phpunit: Remove unnecessary uses of `@before`, `@after`, etc (#42706)
Back when we still supported PHP <=7.1, we had to add these in a lot of
places to work around the fact that PHPUnit 8 started declaring a `void`
return type on `setUp()`, `tearDown()`, and so on, which 7.1 and earler
don't support.

But we've since dropped support for PHP <7.2, so we can swtich back to
the normal functions. It's more straightforward to do that than to start
adding attributes (`#[Before]` and such) for PHPUnit 11+ where the
annotations are deprecated.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/14085626545

Upstream-Ref: Automattic/jetpack@617a716856cc1f098554c281bb51f670814aaecf